+++ src/templates.ts
@@ -1,9 +1,26 @@
-export class Component {
+import * as vscode from 'vscode'
+
+class Base {
+	static getSeparator() {
+		const config = vscode.workspace.getConfiguration('reactTypeScriptToolbox')
+		const indent = config.get<string>('indentation', 'spaces')
+
+		if (indent === "spaces") {
+			return "    "
+		} else {
+			return "\t"
+		}
+	}
+}
+
+export class Component extends Base {
 	static create(name: string, css: string): string {
+		const s = Component.getSeparator()
+
 		let result = '';
 		result += `import * as React from "react"\n`;
 		result += `\n`;
-		
+
 		if (css !== 'none') {
 			result += `/** Stylesheet Imports */\n`;
 			result += `import "./${name}.${css}"\n`;
@@ -11,7 +28,7 @@ export class Component {
 		}
 
 		result += `export interface Props {\n`;
-		result += `\tchildren?: React.ReactNode;\n`;
+		result += `${s}children?: React.ReactNode\n`;
 		result += `}\n`;
 		result += `\n`;
 		result += `export interface State {\n`;
@@ -19,18 +36,17 @@ export class Component {
 		result += `\n`;
 		result += `export default class ${name} extends React.Component<Props, State> {\n`;
 		result += `\n`;
-		// result += `\tpublic state: State\n\n
-		result += `\tconstructor(props: Props) {\n`;
-		result += `\t\tsuper(props)\n\n`;
-		result += `\t\tthis.state = {\n`;
-		result += `\t\t}\n`;
-		result += `\t}\n`;
+		result += `${s}constructor(props: Props) {\n`;
+		result += `${s}${s}super(props)\n\n`;
+		result += `${s}${s}this.state = {\n`;
+		result += `${s}${s}}\n`;
+		result += `${s}}\n`;
 		result += `\n`;
-		result += `\trender() {\n`;
-		result += `\t\treturn (\n`;
-		result += `\t\t\t<div>{ this.props.children }</div>\n`;
-		result += `\t\t)\n`;
-		result += `\t}\n`;
+		result += `${s}render() {\n`;
+		result += `${s}${s}return (\n`;
+		result += `${s}${s}${s}<div>{ this.props.children }</div>\n`;
+		result += `${s}${s})\n`;
+		result += `${s}}\n`;
 		result += `}\n`;
 		return result;
 	}
@@ -46,12 +62,27 @@ export class ExportIndex {
 	}
 }
 
-export class Enum {
+export class Model extends Base {
 	static create(name: string) {
-		let r = ''
+		const s = Component.getSeparator()
 
+		let r = ''
+		r += `export default class ${name} {\n`
+		r += `${s}constructor() {\n`
+		r += `${s}}\n`
+		r += `}`
+
+		return r
+	}
+}
+
+export class Enum extends Base {
+	static create(name: string) {
+		const s = Component.getSeparator()
+
+		let r = ''
 		r += `enum ${name} {\n`
-		r += `\tMyConstant = "MyConstantValue"\n`
+		r += `${s}MyConstant = "MyConstantValue"\n`
 		r += `}\n`
 		r += `export default ${name}`
 
@@ -59,9 +90,11 @@ export class Enum {
 	}
 }
 
-export class Test {
+export class Test extends Base {
 	static create(name: string): string {
-		let result = '';
+		const s = Component.getSeparator()
+
+		let result = ''
 		result += `/** Import React */\n`
 		result += `import * as ReactDOM from 'react-dom';\n`
 		result += `import * as React from 'react';\n`
@@ -76,22 +109,22 @@ export class Test {
 		result += `const classUnderTest = ${name};\n`
 		result += `\n`
 		result += `it('renders without crashing', () => {\n`
-		result += `\tconst div = document.createElement('div');\n`
-		result += `\tReactDOM.render(<${name} />, div);\n`
+		result += `${s}const div = document.createElement('div');\n`
+		result += `${s}ReactDOM.render(<${name} />, div);\n`
 		result += `});\n`
 		result += `\n`
 		result += `describe('render', () => {\n`
 		result += `\n`
-		result += `\tit('should render with props', () => {\n`
-		result += `\t\t// preparation\n`
-		result += `\t\tconst comp = new classUnderTest({ });\n`
-		result += `\t\t\n`
-		result += `\t\t// execution\n`
-		result += `\t\tconst html = shallow(comp.render());\n`
-		result += `\t\t\n`
-		result += `\t\t// testing\n`
-		result += `\t\texpect(html).toContainReact(<div />);\n`
-		result += `\t});\n`
+		result += `${s}it('should render with props', () => {\n`
+		result += `${s}${s}// preparation\n`
+		result += `${s}${s}const comp = new classUnderTest({ });\n`
+		result += `${s}${s}\n`
+		result += `${s}${s}// execution\n`
+		result += `${s}${s}const html = shallow(comp.render());\n`
+		result += `${s}${s}\n`
+		result += `${s}${s}// testing\n`
+		result += `${s}${s}expect(html).toContainReact(<div />);\n`
+		result += `${s}});\n`
 		result += `});\n`
 		return result;
 	}