+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 
 All notable changes to the "react-typescript-toolbox" extension will be documented in this file.
 
+## 1.1.1 - 22.02.2018
+
+### Added
+
+- Support the preferences for Tabs/Spaces as indent
+- Implement Generate Model Class
+- Update Readme
+
 ## 0.10.2 - 31.01.2018
 
 ### Fixed [PathIssue](https://github.com/Sly321/react-typescript-toolbox/issues/3) reported by [jimfilippou](https://github.com/jimfilippou)